Adjustment to how graph is created.
This patch changes the order how we connect and delete reads/writes.
First we connect by line number, since these nodes are more likely to
influence read nodes to written nodes.

Before the patch, nodes with the same name get connected and already
deleted of our tracking dictionaries: reads and writes. The problem with
that is, that trying to connect nodes by line number afterwards will not
happen anymore. That means, that variables written in later line numbers
are not connected in the graph anymore.

Therefore this will hopefully fix #1