Fixes not overwriting double occurrence
This fixes multiple occurrences. The former algorithm simply overwrote
the last write occurrence with the new one, basically removing any
connections of possible initialisation variables, like:

    a = 1
    b = 1
    a = a + b

Starting a slice from the first line would have resulted in nothing,
which is wrong. That happened because the visitor visits from right to
left, overwriting the assigned 'a' from line 1 with the assigned 'a'
from line 3.