cleanup: don't use the IR
I've realised that going through an intermediate representation to
create a control flow graph is not going to work. We need a control flow
graph of exactly the same AST in order to show the graph to the client
how the control flow goes of the source code. If I reduce that into an
IR, we will loose all this information.

It will make sense for a compiler obviously (since the IR then is
transformed further), but not for the static analysis case we're dealing
with.