cfg: include the IdLabelMap
I need this to investigate how this mapping can be used to lookup
identifiers mapped to labels for the CFG. Part of the slice result is
needed to be shown in the editor which variables are dependent. If I
loose this information with all labels, constructing the CFG would be a
meaningless exercise.