cfg: Fixes bug when constructing graph
I think I've made a mistake before resulting in a graph with entry,
middle and exit per line in the source code. That's wrong, since control
flow can fall through with a set of assignments.

This patch fixes this problem, but may need to have changes with support
for branches and conditionals.