cfg: Fuse graphs
The previous code returned a list of control flow graphs according to
the amount of functions we found. What I want tho is one control flow
graph for the entire module. Even though there might not be control flow
going from one function to the other, it still means that we should be
able to represent the module as a graph. Difference is: no edges from
one function to the other.

This patch fuses the closed graphs to one bigger graph.